+++ src/v2019C1/extensions/template/transaction/fsd/extract-to-fsd.ts
@@ -48,6 +48,7 @@ export async function extractToFsd(
 		functionRef,
 		targetParentRef: substationRef,
 		exclude: FSD_EXCLUDE,
+		stripRootAttributes: ['templateUuid'],
 	})
 
 	await cloneFunctionCategories(tx, {