+++ src/v2019C1/extensions/template/transaction/shared/template.types.ts
@@ -1,7 +1,0 @@
-import { Scl } from '@/v2019C1/config'
-
-export type TemplateStructure = {
-	Substation: Scl.RawRecord<'Substation'>
-	VoltageLevel: Scl.RawRecord<'VoltageLevel'>
-	Bay: Scl.RawRecord<'Bay'>
-}